Create P1
so this file is actually just a copy of lab 2 thats been edited. 
he will be releasing the answers for the homework later today, and that will definitely help us so we can replace the motion history with that. 
we can talk more about this later today. 
it mostly runs with no errors.